fix: 과제 기본 utils -> addEvent main.jsx로 재이동 - 멘토링 때 기존 main.jsx에 있는 addEvent는 유지하는 것 말씀 주심!
