+++ src/App.jsx
@@ -7,18 +7,12 @@ export const App = ({ targetPage }) => {
 	const PageComponent = targetPage ?? NotFoundPage;
 	const error = globalStore.getState().error;
 
-	const handleError = (e) => {
-		e.preventDefault();
-		globalStore.setState({ error: null });
-	};
-
 	return (
 		<div>
 			<PageComponent />
 			{error ? (
 				<div
 					id="error-boundary"
-					onClick={handleError}
 					className="fixed bottom-4 left-4 right-4 bg-red-100 border border-red-400 text-red-700 px-4 py-3 rounded-lg shadow-lg transition-opacity duration-300 hover:opacity-75"
 					role="alert"
 				>
